requirements.txt
grpcio-tools
protobuf
opencv-python